requirements.txt
jupyter
ipykernel
webdriver-manager
spotipy
python-dotenv
selenium
pandas
numpy
fuzzywuzzy
beautifulsoup4
requests
python-Levenshtein
matplotlib
seaborn
reportlab
pillow
python-telegram-bot